Hack in fn+hjkl sending arrow key events
This is ultimately _not_ what I want; fn+hjkl is not very ergonomic, and
a far cry from the ultimate s+d simultaneous key press for vi-mode.

But, this is simply dipping my toes into the project to see if I could at
the very least hardcode my setup.

The next step will be trying to get my desired setup hardcoded, which
is:

- caps lock acts as a left_control modifier when pressed with another
  key, but escape when pressed by itself.
- when simultaneously pressing s and d, they acts as a modifier to then
  map hjkl to arrow keys, i.e.,
    - s+d h maps to ←
    - s+d j maps to ↓
    - s+d k maps to ↑
    - s+d l maps to →

Original hack via:
https://github.com/watsoncj/Karabiner-Elements/commit/5531032a309fe99e7e153cb725b9f44a19da69c0